+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Update specification filter to get slug from link.
 
+### Fixed
+- Error on normalize function of the product summary.
+
 ## [0.6.4] - 2018-08-08
 ### Added
 - PriceRange filter.